:sparkles:chore:linear search algo

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,14 +13,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL)
 		return (-1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
